requirements.txt
streamlit
openai
requests
python-dotenv 